requirements.txt
PyQt6
requests
paramiko
telnetlib3
pyssdp
scapy
netifaces
pycryptodome
SQLAlchemy
